requirements.txt
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
flask==2.3.3
Flask-Login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing-extensions==4.8.0
werkzeug==2.3.7
wtforms==3.1.0
zipp==3.17.0